requirements.txt
Flask
google-genai
nltk